+++ frontend/constants.ts
@@ -1,0 +1,8 @@
+// it wasn't clear to me how to easily parse reciept events so I ended up using the topic hashes I saw in logs to determine if the user won or lost
+export const LOSER_TOPIC_HASH =
+  "0x190d4f1249f2e21343a1c0910c21ea36d8a8686a1596352ecc15c4d26a988dcd";
+export const WINNER_TOPIC_HASH =
+  "0x9c2270628a9b29d30ae96b6c4c14ed646ee134febdce38a5b77f2bde9cea2e20";
+
+export const ZKS_SYNC_GOERLI_TESTNET = 280;
+export const CASINO_GAME_ADDRESS = "0xAcDc11Df900624F20A7Fbe85c58cf867C08c279e";
